+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.6 (July 17, 2017)
+
+- Refactor directoryPath to be set from the requiredFlags check.
+
+## 0.0.5 (July 17, 2017)
+
+- Add directory implementation to pull down a list of directories that exist within a given path.
+
 ## 0.0.4 (July 17, 2017)
 
 - Add go-git vendor package which will be used to initiate git commands.